pyproject.toml
[tool.poetry]
name = "api-contacts"
version = "0.1.0"
description = "API for saving contact information"
authors = ["Olesia0402 <olesyashevchuk0402@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12"
libgravatar = "^1.0.4"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.9"
bcrypt = "^4.1.3"
fastapi-mail = "^1.4.1"
python-dotenv = "^1.0.1"
fastapi-limiter = "^0.1.6"
aioredis = "^2.0.1"
cloudinary = "^1.40.0"
pydantic-settings = "^2.2.1"
asyncpg = "^0.29.0"
aiosqlite = "^0.20.0"
fastapi-lifespan-manager = "^0.1.4"


[tool.poetry.group.dev.dependencies]
sphinx = "^7.3.7"


[tool.poetry.group.tests.dependencies]
pytest = "^8.2.2"
pytest-cov = "^5.0.0"
pytest-asyncio = "^0.23.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
